Improve section task handling
Currently it is quite easy to have two tasts in the same
section with a next label, so let's remove the next label
from all cards and recalculate whether the task should
have the next label on every run.

Note this will not increate the API usage, as only the
final diff will be sent to Todoist.

Signed-off-by: Christopher Obbard <chris@64studio.com>